+++ src/components/CmdBarGroup.vue
@@ -1,66 +1,30 @@
 <script setup lang="ts">
-import { useCmdBarEvent } from '../useCmdBarEvent'
+import CmdBarItems from '@/Users/fabiankirchhoff/code/cmd-bar/src/components/CmdBarItems.vue'
 import type { Command, Group } from '../types'
-import { useCmdBarState } from '../useCmdBarState'
-import { computed } from 'vue'
 
-const props = defineProps<{
+defineProps<{
   group: Group
 }>()
 
-const { emitter } = useCmdBarEvent()
-
-const emit = defineEmits<{
+defineEmits<{
   selected: [command: Command]
   clicked: [command: Command]
 }>()
 
-// causes type error!?!?!?
-// defineSlots<{
-//   default(props: { command: Command }): any
-//   loading(props: { group: Group }): any
-// }>()
-
-const isSelectedItem = computed(() => {
-  return (command: Command) => {
-    const isSelected = useCmdBarState?.state.selectedCommandId === command.id
-    if (isSelected) {
-      emitter.emit('selected', command as Command)
-      return true
-    }
-    return false
-  }
-})
-
-const groupIsLoading = computed(() => {
-  return useCmdBarState?.state.groupLoadingStates[props.group.key]
-})
-
-function handleClick(clickedCommand: Command) {
-  emitter.emit('clicked', clickedCommand as Command)
-  useCmdBarState?.executeCommand()
-}
+defineSlots<{
+  default(props: { command: Command }): any
+}>()
 </script>
 
 <template>
   <div class="group-container">
-    <ul v-if="!groupIsLoading && group.commands" data-cmd-bar-items class="items">
-      <li
-        v-for="(command, index) of group.commands"
-        :key="`${group.key}-${index}`"
-        :data-id="command.id"
-        role="option"
-        class="item"
-        :aria-selected="isSelectedItem(command)"
-        @mousemove="useCmdBarState?.selectCommand(command.id)"
-        @click="handleClick(command)"
-      >
-        <slot :command="command" />
-      </li>
+    <ul v-if="group.commands" data-cmd-bar-items class="items">
+      <CmdBarItems v-bind="$attrs" :commands="group.commands">
+        <template #default="{ command }">
+          <slot :command="command" />
+        </template>
+      </CmdBarItems>
     </ul>
-    <div v-else class="loading">
-      <slot name="loading" :group="group" />
-    </div>
   </div>
 </template>