+++ src/composables/useCmdBarEvent.ts
@@ -1,9 +1,11 @@
 import mitt, { Emitter } from 'mitt'
-import { Command } from './types'
+import { Command } from '../types'
 
 type Events = {
   selected: Command
   clicked: Command
+  input: string
+  filterChange: (string | null)[]
 }
 
 const emitter: Emitter<Events> = mitt<Events>()