+++ src/composables/useDefineCommand.ts
@@ -1,4 +1,4 @@
-import type { Command } from './types'
+import type { Command } from '../types'
 
 export function defineCommand(params: Command): Command {
   return params as Command